requirements.txt
asgiref==3.3.4
Django==3.2
django-crispy-forms==1.11.2
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
gunicorn==20.1.0
Markdown==3.3.4
Pillow==8.2.0
psycopg2-binary==2.8.6
PyJWT==2.1.0
python-decouple==3.4
pytz==2021.1
regex==2021.4.4
sqlparse==0.4.1
